backend/requirements.txt
django>=5.0,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.0
django-cors-headers>=4.4
psycopg2-binary>=2.9
python-decouple>=3.8
gunicorn>=22.0
drf-spectacular>=0.27
Pillow>=10.0
